requirements.txt
# Discord Bot with LangChain, LangMem, and MCP Server Tools
discord.py>=2.0.0
python-dotenv>=0.21.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-mistralai>=0.0.1
langchain-google-genai>=0.0.3
langchain-mcp-adapters>=0.0.1
langmem>=0.0.1
langgraph>=0.0.20
chromadb>=0.4.22  # Added for persistent memory
langgraph-checkpoint-sqlite
aiosqlite>=0.20.0
# Core MCP Server library (includes FastAPI, Uvicorn for SSE)
mcp[cli]

# Web scraping and extraction library
crawl4ai

# For loading .env file (API Keys)
python-dotenv

# Potentially needed by crawl4ai or underlying libraries, good to include
# httpx >= 0.27.0 # Usually installed by crawl4ai/mcp
# beautifulsoup4 >= 4.12.2 # Usually installed by crawl4ai

# RSS Feed parsing
feedparser>=6.0.0
beautifulsoup4>=4.12.0